Make MoveCellWise work w/o active tp session
The problem was that DefaultVars throws error unless it finds
an active tp session. This is good for communicating with tp.
But communication is not needed just to move the cursor around.
Therefore, make optional the part about finding the tp session,
which also means that it must be explicitly activated in RunPythonCell.